+++ wls-gui-wahllokalsystem/src/components/dse/stimmzettelerfassung/baseComponents/BaseErfassteStimmzettelCard.vue
@@ -1,56 +1,35 @@
-<script setup lang="ts">
-</script>
 <template>
   <v-card>
     <v-card-title>Erfasste Stimmzettel</v-card-title>
     <v-card-text>
       <v-table>
         <thead>
-        <tr>
-          <th />
-          <th class="font-weight-bold text-right">Insgesamt</th>
-        </tr>
+          <tr>
+            <th />
+            <th class="font-weight-bold text-right">Insgesamt</th>
+          </tr>
         </thead>
         <tbody>
           <tr>
-            <td>
-            Stimmzettel für Beschluss vorgemerkt
-            </td>
-            <td class="text-right">
-              0
-            </td>
+            <td>Stimmzettel für Beschluss vorgemerkt</td>
+            <td class="text-right">0</td>
           </tr>
           <tr>
-            <td>
-              Ungültige Stimmzettel
-            </td>
-            <td class="text-right">
-              0
-            </td>
+            <td>Ungültige Stimmzettel</td>
+            <td class="text-right">0</td>
           </tr>
           <tr>
-            <td>
-              Gültige Stimmzettel
-            </td>
-            <td class="text-right">
-              0
-            </td>
+            <td>Gültige Stimmzettel</td>
+            <td class="text-right">0</td>
           </tr>
         </tbody>
         <tfoot>
           <tr>
-            <td class="font-weight-bold">
-              Stimmzettel Gesamt
-            </td>
-            <td class="font-weight-bold text-right">
-              0
-            </td>
+            <td class="font-weight-bold">Stimmzettel Gesamt</td>
+            <td class="font-weight-bold text-right">0</td>
           </tr>
         </tfoot>
       </v-table>
     </v-card-text>
   </v-card>
 </template>
-
-<script setup lang="ts">
-</script>
